refactor: move MaybeDisjoint cast impls to cast::sets
