+++ intervalsets-core/src/cast/sets.rs
@@ -12,10 +12,13 @@ use num_traits::Bounded;
 
 use super::{Cast, CastElement, LossyCast, LossyCastElement, TryCast, TryCastElement};
 use crate::bound::{BoundType, FiniteBound};
+use crate::disjoint::MaybeDisjoint;
 use crate::error::Error;
 use crate::factory::TrySatisfyFiniteInterval;
 use crate::numeric::Element;
+use crate::ops::{Connects, MergeConnected};
 use crate::sets::{EnumInterval, FiniteInterval, HalfInterval};
+use crate::MaybeEmpty;
 
 // =====================================================================
 // FiniteBound
@@ -251,6 +254,107 @@ where
 }
 
 // =====================================================================
+// MaybeDisjoint — per-variant delegate to the inner `EnumInterval`
+// casts. `Cast` (widening) preserves invariants by definition.
+// `TryCast` errors on cast-induced invariant breakage (mirrors
+// `IntervalSet::try_cast`). `LossyCast` repairs: empties drop,
+// connecting pieces merge — consistent with element-layer distinctions
+// already discarded.
+// =====================================================================
+
+impl<T, U> Cast<MaybeDisjoint<U>> for MaybeDisjoint<T>
+where
+    T: CastElement<U>,
+    U: Element,
+{
+    type Output = MaybeDisjoint<U>;
+
+    fn cast(self) -> MaybeDisjoint<U> {
+        match self {
+            MaybeDisjoint::Consumed => MaybeDisjoint::Consumed,
+            MaybeDisjoint::Connected(i) => MaybeDisjoint::Connected(i.cast()),
+            // Monotone widening preserves non-empty, ordering, and
+            // non-connecting invariants of `Disjoint`.
+            MaybeDisjoint::Disjoint(a, b) => MaybeDisjoint::Disjoint(a.cast(), b.cast()),
+        }
+    }
+}
+
+impl<T, U> TryCast<MaybeDisjoint<U>> for MaybeDisjoint<T>
+where
+    T: TryCastElement<U>,
+    U: Element,
+{
+    type Output = MaybeDisjoint<U>;
+    type Error = Error;
+
+    fn try_cast(self) -> Result<MaybeDisjoint<U>, Error> {
+        match self {
+            MaybeDisjoint::Consumed => Ok(MaybeDisjoint::Consumed),
+            MaybeDisjoint::Connected(i) => {
+                // From<EnumInterval<U>> normalizes empty → Consumed.
+                let cast: EnumInterval<U> = i.try_cast()?;
+                Ok(MaybeDisjoint::from(cast))
+            }
+            MaybeDisjoint::Disjoint(a, b) => {
+                let a: EnumInterval<U> = a.try_cast()?;
+                let b: EnumInterval<U> = b.try_cast()?;
+                // Strict: post-cast intervals must still satisfy the
+                // `Disjoint` invariants (non-empty, sorted, non-touching).
+                // Any narrowing-induced violation surfaces as
+                // `InvalidBoundPair` — the closest existing variant for
+                // "a paired-structure invariant has broken".
+                if a.is_empty() || b.is_empty() || a >= b || a.connects(&b) {
+                    return Err(Error::InvalidBoundPair);
+                }
+                Ok(MaybeDisjoint::Disjoint(a, b))
+            }
+        }
+    }
+}
+
+impl<T, U> LossyCast<MaybeDisjoint<U>> for MaybeDisjoint<T>
+where
+    T: LossyCastElement<U>,
+    U: Element + Bounded,
+{
+    type Output = MaybeDisjoint<U>;
+
+    fn lossy_cast(self) -> MaybeDisjoint<U> {
+        match self {
+            MaybeDisjoint::Consumed => MaybeDisjoint::Consumed,
+            MaybeDisjoint::Connected(i) => MaybeDisjoint::from(i.lossy_cast()),
+            MaybeDisjoint::Disjoint(a, b) => {
+                let a: EnumInterval<U> = a.lossy_cast();
+                let b: EnumInterval<U> = b.lossy_cast();
+
+                // Repair: drop empties first.
+                match (a.is_empty(), b.is_empty()) {
+                    (true, true) => MaybeDisjoint::Consumed,
+                    (true, false) => MaybeDisjoint::Connected(b),
+                    (false, true) => MaybeDisjoint::Connected(a),
+                    (false, false) => {
+                        // Narrowing can flip relative order at the extremes
+                        // (e.g. both saturate the same direction). Re-sort.
+                        let (lo, hi) = if a <= b { (a, b) } else { (b, a) };
+                        if lo.connects(&hi) {
+                            match lo.merge_connected(hi) {
+                                Some(merged) => MaybeDisjoint::Connected(merged),
+                                None => {
+                                    unreachable!("connects() implies merge_connected returns Some")
+                                }
+                            }
+                        } else {
+                            MaybeDisjoint::Disjoint(lo, hi)
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+// =====================================================================
 // Tests
 // =====================================================================
 
@@ -434,5 +538,109 @@ mod tests {
     fn nan_via_tier4_bypass_panics_in_lossy_cast_to_int() {
         let bad = FiniteBound::new_assume_valid(BoundType::Closed, f64::NAN);
         let _: FiniteBound<i32> = bad.lossy_cast();
+    }
+
+    // ---------- MaybeDisjoint ----------
+
+    #[test]
+    fn cast_maybe_disjoint_consumed() {
+        let x = MaybeDisjoint::<i32>::Consumed;
+        let y: MaybeDisjoint<i64> = x.cast();
+        assert!(matches!(y, MaybeDisjoint::Consumed));
+    }
+
+    #[test]
+    fn cast_maybe_disjoint_connected_widening() {
+        let x = MaybeDisjoint::Connected(EnumInterval::closed(0_i32, 10));
+        let y: MaybeDisjoint<i64> = x.cast();
+        match y {
+            MaybeDisjoint::Connected(i) => assert_eq!(i, EnumInterval::closed(0_i64, 10)),
+            _ => panic!("expected Connected"),
+        }
+    }
+
+    #[test]
+    fn cast_maybe_disjoint_disjoint_widening_preserves_invariants() {
+        let x = MaybeDisjoint::Disjoint(
+            EnumInterval::closed(0_i32, 10),
+            EnumInterval::closed(20_i32, 30),
+        );
+        let y: MaybeDisjoint<i64> = x.cast();
+        match y {
+            MaybeDisjoint::Disjoint(a, b) => {
+                assert_eq!(a, EnumInterval::closed(0_i64, 10));
+                assert_eq!(b, EnumInterval::closed(20_i64, 30));
+            }
+            _ => panic!("expected Disjoint"),
+        }
+    }
+
+    #[test]
+    fn try_cast_maybe_disjoint_widening() {
+        let x = MaybeDisjoint::Disjoint(
+            EnumInterval::closed(0_i32, 10),
+            EnumInterval::closed(20_i32, 30),
+        );
+        let y: MaybeDisjoint<i64> = x.try_cast().unwrap();
+        assert!(matches!(y, MaybeDisjoint::Disjoint(_, _)));
+    }
+
+    #[test]
+    fn try_cast_maybe_disjoint_collision_errors() {
+        // Two f64 intervals whose f32 projections collide → narrowing
+        // breaks Disjoint invariants → InvalidBoundPair.
+        let lo1 = 1.0_f64;
+        let hi1 = 1.0_f64 + 4.0 * f64::EPSILON;
+        let lo2 = 1.0_f64 + 5.0 * f64::EPSILON;
+        let hi2 = 1.0_f64 + 9.0 * f64::EPSILON;
+        assert_eq!(lo1 as f32, lo2 as f32);
+        let x = MaybeDisjoint::Disjoint(
+            EnumInterval::Finite(
+                FiniteInterval::try_new(FiniteBound::closed(lo1), FiniteBound::closed(hi1))
+                    .unwrap(),
+            ),
+            EnumInterval::Finite(
+                FiniteInterval::try_new(FiniteBound::closed(lo2), FiniteBound::closed(hi2))
+                    .unwrap(),
+            ),
+        );
+        let y: Result<MaybeDisjoint<f32>, _> = x.try_cast();
+        assert!(matches!(y, Err(Error::InvalidBoundPair)));
+    }
+
+    #[test]
+    fn lossy_cast_maybe_disjoint_collision_merges() {
+        let lo1 = 1.0_f64;
+        let hi1 = 1.0_f64 + 4.0 * f64::EPSILON;
+        let lo2 = 1.0_f64 + 5.0 * f64::EPSILON;
+        let hi2 = 1.0_f64 + 9.0 * f64::EPSILON;
+        let x = MaybeDisjoint::Disjoint(
+            EnumInterval::Finite(
+                FiniteInterval::try_new(FiniteBound::closed(lo1), FiniteBound::closed(hi1))
+                    .unwrap(),
+            ),
+            EnumInterval::Finite(
+                FiniteInterval::try_new(FiniteBound::closed(lo2), FiniteBound::closed(hi2))
+                    .unwrap(),
+            ),
+        );
+        let y: MaybeDisjoint<f32> = x.lossy_cast();
+        assert!(!matches!(y, MaybeDisjoint::Disjoint(_, _)));
+    }
+
+    #[test]
+    fn lossy_cast_maybe_disjoint_widening_preserves_invariants() {
+        let x = MaybeDisjoint::Disjoint(
+            EnumInterval::closed(0_i64, 10),
+            EnumInterval::closed(20_i64, 30),
+        );
+        let y: MaybeDisjoint<i32> = x.lossy_cast();
+        match y {
+            MaybeDisjoint::Disjoint(a, b) => {
+                assert_eq!(a, EnumInterval::closed(0_i32, 10));
+                assert_eq!(b, EnumInterval::closed(20_i32, 30));
+            }
+            _ => panic!("expected Disjoint after lossless widening"),
+        }
     }
 }
